backend/requirements.txt
fastapi==0.116.1
uvicorn==0.32.1
motor==3.6.0
pydantic==2.10.4
python-dotenv==1.0.1
starlette==0.47.2
email-validator==2.2.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
bcrypt==3.2.0
requests==2.32.4
gunicorn
slowapi
async-lru
beanie==1.26.0
pytest
pytest-asyncio
mongomock-motor
httpx
python-multipart
pydantic-settings==2.1.0
tenacity==8.2.3
fastapi-cache2[redis]
sentry-sdk[fastapi]==2.18.0
structlog==24.1.0
SQLAlchemy==2.0.23
locust==2.24.0

pytest
pytest-asyncio
pytest-mock
mongomock-motor
beanie
pydantic[email]
httpx
fastapi
passlib
slowapi
fastapi-cache2
pydantic-settings
redis
python-multipart
tenacity
gunicorn
